Move Model\Item::getPlaintextPost to Model\ItemContent::getPlaintextPost
